# services/server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Leasy Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/lecture": {
            "get": {
                "tags": [
                    "Lectures"
                ],
                "description": "Retrieve lecture records. Supports pagination and sorting.",
                "operationId": "getLectures",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "only_favorites",
                        "in": "query",
                        "description": "Filter only favorite lectures (true or false).",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "search_by_title",
                        "in": "query",
                        "description": "Search lectures by title prefix (e.g., 'Intro' matches 'Introduction to Physics').",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Intro"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by column (e.g., 'date', 'name', 'progress').",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "date",
                                "name",
                                "progress"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction (e.g., 'asc', 'desc').",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with lectures data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            },
            "post": {
                "tags": [
                    "Lectures"
                ],
                "description": "Store a lecture with the provided details",
                "operationId": "postLecture",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "video",
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "video": {
                                        "description": "The video file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "title": {
                                        "description": "The title of the lecture",
                                        "type": "string",
                                        "example": "Introduction to Physics"
                                    },
                                    "description": {
                                        "description": "The description of the lecture",
                                        "type": "string",
                                        "example": "An introductory lecture on the fundamentals of physics."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lecture created successfully"
                    },
                    "500": {
                        "description": "An error occurred"
                    },
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/lecture/{uuid}": {
            "get": {
                "tags": [
                    "Lectures"
                ],
                "description": "Fetch detailed information about a specific lecture using its UUID.",
                "operationId": "getLecture",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the lecture",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with lecture data"
                    },
                    "404": {
                        "description": "Lecture not found"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/lecture/add-to-favorites/{uuid}": {
            "put": {
                "tags": [
                    "Lectures"
                ],
                "description": "Add a lecture to the user's favorites.",
                "operationId": "addLectureToFavorites",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the lecture to add to favorites.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lecture has been successfully added to favorites"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Lecture Not Found"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
...
    }
}